% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome_features.R
\name{biome_features}
\alias{biome_features}
\title{Obtain Biome Features from Occurrence Records}
\usage{
biome_features(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  biome.input = NULL,
  biome.id = "BIOME",
  download.path = NULL
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{biome.input}{s simple features collection of geometry type polygon, contain polygons of different biomes.
If NULL, the WWF biome scheme is downloaded from
https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}

\item{biome.id}{a character string. The name of the column with the biome names in biome.input.
Default is "BIOME"}

\item{download.path}{character string. The path were to save the WWF polygons
if biome.input is NULL. Default is the working directory}
}
\value{
a data.frame of climatic features
}
\description{
Will code all species in the input file into biomes based on an
intersection of the coordinates with a shape file. The biome scheme can
be user provided or by default will download the WWF biomes.
}
\details{
If biome.input is NULL this will download  the WWF biome scheme from
https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world
and save them in the working directory
}
\examples{
\dontrun{
dat <- data.frame(species = "A",
                decimallongitude = runif (200,-5,5),
                decimallatitude = runif (200,-5,5))

biom_features(dat)
}


}
\seealso{
Other Feature preparation: 
\code{\link{clim_features}()},
\code{\link{geo_features}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
