% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npbnn_helper_functions.R
\name{bnn_load_data}
\alias{bnn_load_data}
\title{npBNN helper functions (these are run internally and don't require a documentation)}
\usage{
bnn_load_data(
  features,
  labels,
  seed = 1234,
  testsize = 0.1,
  all_class_in_testset = TRUE,
  header = TRUE,
  instance_id = TRUE,
  from_file = FALSE
)
}
\description{
npBNN helper functions (these are run internally and don't require a documentation)
}
