% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_test_accuracy.R
\name{calculate_test_accuracy}
\alias{calculate_test_accuracy}
\title{Calculate test accuracy}
\usage{
calculate_test_accuracy(bnn_data, logger, bnn_model, post_summary_mode = 0)
}
\description{
Calculate test accuracy
}
