% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_BNN_model.R
\name{create_BNN_model}
\alias{create_BNN_model}
\title{Set up BNN model structure}
\usage{
create_BNN_model(
  feature_data,
  n_nodes_list,
  seed = 1234,
  use_class_weight = TRUE,
  use_bias_node = TRUE,
  prior = 1,
  p_scale = 1,
  init_std = 0.1
)
}
\description{
Set up BNN model structure
}
