% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{feature_importance}
\alias{feature_importance}
\title{Determine feature importance}
\usage{
feature_importance(
  bnn_data,
  logger,
  bnn_model,
  n_permutations = 100,
  predictions_outdir = "feature_importance",
  feature_blocks = list(c(0, 1, 2, 3, 4, 5, 6, 7), c(8, 9, 10), c(11, 12, 13, 14, 15,
    16, 17, 18, 19, 20)),
  unlink_features_within_block = TRUE
)
}
\description{
Determine feature importance
}
