% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_features.R
\name{footprint_features}
\alias{footprint_features}
\title{Extract Human Footprint Index Features from Occurrence Records}
\source{
https://wcshumanfootprint.org/
}
\usage{
footprint_features(
  x,
  footp_input = NULL,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  rescale = TRUE,
  year = c(1993, 2009),
  file_path = NULL,
  breaks = c(0, 0.81, 1.6, 2.3, 100)
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{footp_input}{an object of the class raster or RasterStack with values for the human footprint index.
If a RasterStack, different layers are interpreted as different time-slices.}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{rescale}{logical. If TRUE, the values are rescaled using natural logarithm transformation. If FALSE,
remember to change the breaks argument.}

\item{year}{numeric. The years for which to obtain the human footprint index.
The default is to the two layers available. Can be a either year, in case only
one slice is desired. Other time slices are currently not supported,}

\item{file_path}{a character string. The path where raster can be saved on disk.
IF NULL the working directory. Default = NULL.}

\item{breaks}{numerical. The breaks to bin the human footprint index for the final features. The defaults are
empirical values for the global footprint and rescale=TRUE. For custom values ensure that they
cover the whole value range and are adapted to the value of rescale.}
}
\value{
a data.frame of human footprint features
}
\description{
Bins the human footprint index into a set of bins and the fraction of occurrence records
of a species in each bin are the features. By default the human footprint index is downloaded from
https://wcshumanfootprint.org/. THIS FUNCTION WILL DOWNLOAD DATA FROM
THE INTERNET AND SAVE IT TO THE  WORKING DIRECTORY. The data files are >200 MB each and downloading may
take some time on first execution.
}
\details{
By default four categories of increasing human footprint index ( 1 = lowest, 4 = highest)
are selected and rescaled.
}
\examples{
\dontrun{
dat <- data.frame(species = "A",
                decimallongitude = runif (200,-5,5),
                decimallatitude = runif (200,-5,5))

footprint_features(dat)
}


}
\seealso{
Other Feature preparation: 
\code{\link{bias_features}()},
\code{\link{biome_features}()},
\code{\link{clim_features}()},
\code{\link{geo_features}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
