% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_bias.R
\name{ft_bias}
\alias{ft_bias}
\title{Extract Bias Features from Occurrence Records}
\usage{
ft_bias(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  res = 0.5,
  ras = NULL,
  plot = TRUE
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{res}{numeric. The resolution of the default resolution to calculate sampling bias. In decimal degrees.}

\item{ras}{a raster object. Alternative to res, a sample raster to calculate sampling bias. Needs to use the same CRS as
the coordinates in x.}

\item{plot}{logical. Should the results of the sampbias analysis be plotted for diagnostics?}
}
\value{
a data.frame of bias features
}
\description{
Use the sampbias method to assess the geographic sampling bias at the locations where a species is collected and the range of
sampling bias for all records per species.Values summarized per species are the median and the 0.05 to 0.95 percentiles.
}
\details{
See the ?sampbias::calculate_bias for details.
}
\examples{
\dontrun{
dat <- data.frame(species = "A",
                decimallongitude = runif (200,-5,5),
                decimallatitude = runif (200,-5,5))

ft_bias(dat)
}


}
\seealso{
Other Feature preparation: 
\code{\link{ft_biom}()},
\code{\link{ft_clim}()},
\code{\link{ft_foot}()},
\code{\link{ft_geo}()},
\code{\link{prep_features}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
