% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_biom.R
\name{ft_biom}
\alias{ft_biom}
\title{Obtain Biome Features from Occurrence Records}
\usage{
ft_biom(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  biome.input = NULL,
  biome.id = "BIOME",
  download.path = NULL,
  remove_zeros = FALSE
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{biome.input}{s simple features collection of geometry type polygon, contain polygons of different biomes.
If NULL, the WWF biome scheme is downloaded from
https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}

\item{biome.id}{a character string. The name of the column with the biome names in biome.input.
Default is "BIOME"}

\item{download.path}{character string. The path were to save the WWF polygons
if biome.input is NULL. Default is the working directory}

\item{remove_zeros}{logical. If TRUE biomes without occurrence of any species are removed from the features.
Default = FALSE}
}
\value{
a data.frame of climatic features
}
\description{
Will code all species in the input file into biomes based on an
intersection of the coordinates with a shape file. The biome scheme can
be user provided or by default will download the WWF biomes.
}
\details{
If biome.input is NULL this will download  the WWF biome scheme from
https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world
and save them in the working directory
}
\examples{
\dontrun{
dat <- data.frame(species = c("A","b"),
                  decimallongitude = runif (200,10,15),
                  decimallatitude = runif (200,-5,5))

ft_biom(dat)
}


}
\seealso{
Other Feature preparation: 
\code{\link{ft_bias}()},
\code{\link{ft_clim}()},
\code{\link{ft_foot}()},
\code{\link{ft_geo}()},
\code{\link{prep_features}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
