% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_geo.R
\name{ft_geo}
\alias{ft_geo}
\title{Extract Geographic Features from Occurrence Records}
\usage{
ft_geo(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  rescale = TRUE
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{rescale}{logical. If TRUE, the geographic features are rescaled.
This is recommended to run IUCNN, and the default. If FALSE, raw (human readable)
feature values are returned.}
}
\value{
a data.frame of geographic features
}
\description{
Calculates the number of occurrences, number of unique occurrences,
mean latitude, mean longitude, latitudinal range, longitudinal range,
eoo, aoo and hemisphere as input features for IUCNN from a list of species occurrences.
}
\details{
Coordinate ranges are 90\% quantiles, for species with less than three occurrences EOO is set to AOO.
}
\examples{
dat <- data.frame(species = c("A","B"),
                  decimallongitude = runif (200,10,15),
                  decimallatitude = runif (200,-5,5))

ft_geo(dat)

}
\seealso{
Other Feature preparation: 
\code{\link{ft_bias}()},
\code{\link{ft_biom}()},
\code{\link{ft_clim}()},
\code{\link{ft_foot}()},
\code{\link{prep_features}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
