% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_features.R
\name{geo_features}
\alias{geo_features}
\title{Extract Geographic Features from Occurrence Records}
\usage{
geo_features(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  rescale = TRUE
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{rescale}{logical. If TRUE, the geographic features are rescaled.
This is recommended to run IUCNN, and the default. If FALSE, raw (human readable)
feature values are returned.}
}
\value{
a data.frame of geographic features
}
\description{
Calculates the number of occurrences, number of unique occurrences,
mean latitude, mean longitude, latitudinal range, longitudinal range,
eoo, aoo and hemisphere as input features for IUCNN from a list of species occurrences.
}
\details{
Coordinate ranges are 90\% quantiles, for species with less than three occurrences EOO is set to AOO.
}
\examples{

dat <- data.frame(species = "A",
                decimallongitude = runif (200,-5,5),
                decimallatitude = runif (200,-5,5))

geo_features(dat)

}
\seealso{
Other Feature preparation: 
\code{\link{bias_features}()},
\code{\link{biome_features}()},
\code{\link{bnn_load_data}()},
\code{\link{clim_features}()},
\code{\link{footprint_features}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
