% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNN-package.R
\docType{package}
\name{IUCNN-package}
\alias{IUCNN-package}
\alias{IUCNN}
\title{Neural Networks to Approximate IUCN Red List Conservation Assessments}
\description{
Train neural networks based on geographic species occurrences, environmental
data and existing IUCN Red List assessments to predict the conservation status
of "Not Evaluated" species, for any taxon or geographic region of interest.
}
\details{
\tabular{ll}{
Package: \tab IUCNN\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
License: \tab LGPL-2.1\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IUCNN/IUCNN}
  \item Report bugs at \url{https://github.com/IUCNN/IUCNN/issues}
}

}
\author{
\strong{Alexander Zizka}\cr
(email: \email{alexander.zizka@idiv.de})

\strong{Tobias Andermann}

\strong{Daniele Silvestro}
}
\keyword{package}
