% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn_bias_features.R
\name{iucnn_bias_features}
\alias{iucnn_bias_features}
\title{Extract Bias Features from Occurrence Records}
\usage{
iucnn_bias_features(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  res = 0.5,
  ras = NULL,
  plot = TRUE
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{res}{numeric. The resolution of the default resolution to calculate sampling bias. In decimal degrees.}

\item{ras}{a SpatRaster object. Alternative to res, a sample SpatRaster to
calculate sampling bias. Needs to use the same CRS as the coordinates in x.}

\item{plot}{logical. Should the results of the sampbias analysis be plotted for diagnostics?}
}
\value{
a data.frame of bias features
}
\description{
Use the sampbias method to assess the geographic sampling bias at the locations where a species is collected and the range of
sampling bias for all records per species.Values summarized per species are the median and the 0.05 to 0.95 percentiles.
}
\details{
See the ?sampbias::calculate_bias for details.
}
\examples{
\dontrun{
dat <- data.frame(species = c("A", "b"),
                  decimallongitude = runif(200, 10, 15),
                  decimallatitude = runif(200, -5, 5))
iucnn_bias_features(dat)

}


}
\seealso{
Other Feature preparation: 
\code{\link{iucnn_biome_features}()},
\code{\link{iucnn_climate_features}()},
\code{\link{iucnn_cnn_features}()},
\code{\link{iucnn_footprint_features}()},
\code{\link{iucnn_geography_features}()},
\code{\link{iucnn_phylogenetic_features}()},
\code{\link{iucnn_prepare_features}()},
\code{\link{iucnn_prepare_phy}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
