% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn_biome_features.R
\name{iucnn_biome_features}
\alias{iucnn_biome_features}
\title{Obtain Biome Features from Occurrence Records}
\usage{
iucnn_biome_features(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  biome_input = NULL,
  biome_id = "BIOME",
  download_folder = tempdir(),
  remove_zeros = FALSE
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{biome_input}{a simple features collection of geometry type polygon,
contain polygons of different biomes.
If NULL, the WWF biome scheme is downloaded from
https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}

\item{biome_id}{a character string. The name of the column
with the biome names in biome_input.
Default is "BIOME"}

\item{download_folder}{character string. The folder were to save the
data used for feature extraction. Relative to the working directory.
Set to NULL for the working directory}

\item{remove_zeros}{logical. If TRUE biomes without occurrence of
any species are removed from the features.
Default = FALSE}
}
\value{
a data.frame of climatic features
}
\description{
Will code all species in the input file into biomes based on an
intersection of the coordinates with a shape file. The biome scheme can
be user provided or by default will download the WWF biomes.
}
\details{
If biome_input is NULL this will download  the WWF biome scheme from
https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world
and save them in the working directory
}
\examples{
\donttest{
dat <- data.frame(species = c("A", "b"),
                  decimallongitude = runif(200, 10, 15),
                  decimallatitude = runif(200, -5, 5))

iucnn_biome_features(dat)
}

}
\seealso{
Other Feature preparation: 
\code{\link{iucnn_bias_features}()},
\code{\link{iucnn_climate_features}()},
\code{\link{iucnn_cnn_features}()},
\code{\link{iucnn_footprint_features}()},
\code{\link{iucnn_geography_features}()},
\code{\link{iucnn_phylogenetic_features}()},
\code{\link{iucnn_prepare_features}()},
\code{\link{iucnn_prepare_phy}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
