% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn_cnn_features.R
\name{iucnn_cnn_features}
\alias{iucnn_cnn_features}
\title{Prepare Features for a CNN model}
\usage{
iucnn_cnn_features(
  x,
  y = NULL,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  crs_x = "+proj=longlat +datum=WGS84",
  res_y = 1
)
}
\arguments{
\item{x}{a data.frame with at least three columns containing taxon name, decimal longitude and latitude values.}

\item{y}{a raster as reference to count the number of occurrence records in.
Can be of any resolution and CRS but the coordinates in x need to be in the same CRS.}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{crs_x}{a proj4string specifying the Coordinate Reference system of the coordinates in x.
Default is to lat/lon WGS84.}

\item{res_y}{numeric. The resolution for the raster in decimal degrees.
Only relevant if y is not provided.}
}
\value{
a list of matrices, one for each input species, where the cells represent the number
of occurrence records in this cell as input for the \dQuote{cnn} class of \code{\link{iucnn_train_model}}.
}
\description{
Converts a data.frame of species occurrences into input features to use
a convolutional neural network to approximate species extinction risk.
}
\details{
If y is not provided, assumes a lat/lon grid with extent equal to the respective minimum and maximum in x
}
\examples{
dat <- data.frame(species = c("A","B"),
                  decimallongitude = runif (200,10,15),
                  decimallatitude = runif (200,-5,5))

iucnn_cnn_features(dat)


}
\seealso{
Other Feature preparation: 
\code{\link{iucnn_bias_features}()},
\code{\link{iucnn_biome_features}()},
\code{\link{iucnn_climate_features}()},
\code{\link{iucnn_footprint_features}()},
\code{\link{iucnn_geography_features}()},
\code{\link{iucnn_phylogenetic_features}()},
\code{\link{iucnn_prepare_features}()},
\code{\link{iucnn_prepare_phy}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
