% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn_geography_features.R
\name{iucnn_geography_features}
\alias{iucnn_geography_features}
\title{Extract Geographic Features from Occurrence Records}
\usage{
iucnn_geography_features(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  rescale = TRUE
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{rescale}{logical. If TRUE, the geographic features are rescaled.
This is recommended to run IUCNN, and the default. If FALSE, raw (human readable)
feature values are returned.}
}
\value{
a data.frame of geographic features
}
\description{
Calculates the number of occurrences, number of unique occurrences,
mean latitude, mean longitude, latitudinal range, longitudinal range,
eoo, aoo and hemisphere as input features for IUCNN
from a list of species occurrences.
}
\details{
Coordinate ranges are 90\% quantiles, for species with
less than three occurrences EOO is set to AOO.
}
\examples{
dat <- data.frame(species = c("A","B"),
                  decimallongitude = runif (200,10,15),
                  decimallatitude = runif (200,-5,5))

iucnn_geography_features(dat)

}
\seealso{
Other Feature preparation: 
\code{\link{iucnn_bias_features}()},
\code{\link{iucnn_biome_features}()},
\code{\link{iucnn_climate_features}()},
\code{\link{iucnn_cnn_features}()},
\code{\link{iucnn_footprint_features}()},
\code{\link{iucnn_phylogenetic_features}()},
\code{\link{iucnn_prepare_features}()},
\code{\link{iucnn_prepare_phy}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
