% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn_phylogenetic_features.R
\name{iucnn_phylogenetic_features}
\alias{iucnn_phylogenetic_features}
\title{Extract Phylogenetic Features Based on Phylogenetic Eigenvectors}
\usage{
iucnn_phylogenetic_features(x, species = "species", phy.eigen)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{phy.eigen}{a matrix of phylogenetic Eigenvector calculated
from a phylogenetic tree including training and prediction species
using the \code{iucnn_prepare_phy} function.}
}
\value{
a data.frame of phylogenetic features
}
\description{
Extract features for a list of species based on phylogenetic Eigenvectors
prepared with the \code{iucnn_prepare_phy} function.
Tip labels need to match the species names in x.
}
\examples{
dat <- data.frame(species = c("A","B", "X"),
                  decimallongitude = runif(180,10,15),
                  decimallatitude = runif(180,-5,5))

tree <- ape::rphylo(n = 10, birth = 0.1, death = 0)
tree$tip.label <- LETTERS[1:10]
phy <- iucnn_prepare_phy(phy = tree)

iucnn_phylogenetic_features(x = dat,
                           phy.eigen = phy)

}
\seealso{
Other Feature preparation: 
\code{\link{iucnn_bias_features}()},
\code{\link{iucnn_biome_features}()},
\code{\link{iucnn_climate_features}()},
\code{\link{iucnn_cnn_features}()},
\code{\link{iucnn_footprint_features}()},
\code{\link{iucnn_geography_features}()},
\code{\link{iucnn_prepare_features}()},
\code{\link{iucnn_prepare_phy}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
