% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn_prepare_features.R
\name{iucnn_prepare_features}
\alias{iucnn_prepare_features}
\title{Prepare features for an IUCNN model}
\usage{
iucnn_prepare_features(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  type = c("geographic", "biomes", "climate", "humanfootprint"),
  download_folder = tempdir(),
  impute_features = TRUE
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{type}{character. The type of features to calculate. Possible options are
\dQuote{geographic}, \dQuote{biome}, \dQuote{climate},
\dQuote{human footprint}.}

\item{download_folder}{character string. The folder were to save the
data used for feature extraction. Relative to the working directory.
Set to NULL for the working directory}

\item{impute_features}{logical. Should NA values be imputed for features? If TRUE
NA values are filled using multiple imputation using missForest. If set to FALSE
species with missing values are removed during the training process.}
}
\value{
a data.frame of features
}
\description{
A wrapper function to prepare all default features included in IUCNN:
geographic, biomes, climate, human footprint.
If desired, bias features need to be calculated separately with ft_bias.
For more control over feature preparation, you can use the
\code{\link{iucnn_geography_features}}, \code{\link{iucnn_biome_features}}, \code{\link{iucnn_climate_features}},
\code{\link{iucnn_footprint_features}} functions.
}
\details{
Without internet access, only geographic features are calculated,
}
\examples{
\donttest{
dat <- data.frame(species = c("A","B"),
                  decimallongitude = runif (200,10,15),
                  decimallatitude = runif (200,-5,5))

iucnn_prepare_features(dat)
}
}
\seealso{
Other Feature preparation: 
\code{\link{iucnn_bias_features}()},
\code{\link{iucnn_biome_features}()},
\code{\link{iucnn_climate_features}()},
\code{\link{iucnn_cnn_features}()},
\code{\link{iucnn_footprint_features}()},
\code{\link{iucnn_geography_features}()},
\code{\link{iucnn_phylogenetic_features}()},
\code{\link{iucnn_prepare_phy}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
