% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labels_broad}
\alias{labels_broad}
\title{Binary Threat Status for 884 orchid species}
\format{
A data frame with 884 rows and 2 variables:
\describe{
\item{species}{The canonical species name}
\item{labels}{The IUCN conservation assessment, converted to numerical for use with IUCNN.
0 =  Threatened, 1 = Not Threatened}
}
}
\source{
\url{https://www.iucnredlist.org/}
}
\usage{
labels_broad
}
\description{
A dataset containing the International Union for the Conservation of Nature's
Global Red List conservation assessments for 884 species of orchids.
Converted to the Binary "threatened/not threatened" level.
}
\keyword{datasets}
