% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_setup.R
\name{MCMC_setup}
\alias{MCMC_setup}
\title{Set up MCMC for BNN}
\usage{
MCMC_setup(
  bnn_model,
  update_f,
  update_ws,
  MCMC_temperature = 1,
  likelihood_tempering = 1,
  n_iteration = 5000,
  sampling_f = 10,
  print_f = 1000,
  n_post_samples = 100,
  sample_from_prior = FALSE
)
}
\description{
Set up MCMC for BNN
}
