% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orchid_classes}
\alias{orchid_classes}
\title{Geographic Occurrence Records for Orchids with IUCN assessment}
\format{
A data frame with 884 rows and 3 variables:
\describe{
\item{species}{The canonical species name}
\item{category_broad}{The IUCN conservation assessment, converted to numerical for use with IUCNN.
Threatened ("CR", "EN", or "VU") Not Threatened ("NT" or "LC")}
\item{category_detail}{The IUCN conservation assessment, converted to numerical for use with IUCNN.
0 =  Critically ENdangered (CR), 1 = Endangered (EN), 2 = Vulnerable (VU), 3 = Near Threatened (NT), 4 = Least Concern (LC)}
}
}
\source{
\url{https://www.iucnredlist.org/}
}
\usage{
orchid_classes
}
\description{
A dataset containing geo-referenced occurrences of 884 Orchid species with
existing IUCN Red List assessment ("CR", "EN", "VU", "NT", "LC"). This is example data
to train an IUCNN model.
}
\keyword{datasets}
