% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orchid_occ}
\alias{orchid_occ}
\title{Geographic Occurrence Records for Orchids with IUCN assessment}
\format{
A data frame with 125,412 rows and 3 variables:
\describe{
\item{species}{The canonical species name}
\item{decimallongitude}{longitudinal coordinates}
\item{decimallatitude}{latitudinal coordinates}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
orchid_occ
}
\description{
A dataset containing geo-referenced occurrences of 884 Orchid species with
existing IUCN Red List assessment ("CR", "EN", "VU", "NT", "LC"). This is example data
to train an IUCNN model.
}
\keyword{datasets}
