% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pdp.R
\name{plot.iucnn_pdp}
\alias{plot.iucnn_pdp}
\title{Plot partial dependence probabilities}
\usage{
\method{plot}{iucnn_pdp}(
  x,
  ask = FALSE,
  features = NULL,
  uncertainty = TRUE,
  order_categorical = TRUE,
  col = NULL,
  ...
)
}
\arguments{
\item{x}{iucnn_get_pdp output. A list of class icunn_pdp named according to
the feature and the partial dependence probabilities for the IUCN category.
See \code{\link{iucnn_get_pdp}}}

\item{ask}{logical (default = FALSE). Indicates if the user is prompted before
a new plot will be displayed.}

\item{features}{(default=NULL). Optional vector of integers specifying which
feature's PD should be plotted, e.g. c(1, 3).}

\item{uncertainty}{logical (default = TRUE). Should dropout uncertainty be
displayed?}

\item{order_categorical}{(default = TRUE). Should categorical features be
ordered so that extinction risk increased along the x-axis or should they be
displayed in the order given by x.}

\item{col}{(default = NULL). Custom colors for IUCNN categories.}

\item{...}{further graphical parameter for par, e.g. different margins with
mar = c(10, 4, 0.5, 0.5)}
}
\description{
Plot partial dependence probabilities
}
