% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prediction_occ}
\alias{prediction_occ}
\title{Geographic Occurrence Records for Not Evaluated Orchids}
\format{
A data frame with 14,900 rows and 3 variables:
\describe{
\item{species}{The canonical species name}
\item{decimallongitude}{longitudinal coordinates}
\item{decimallatitude}{latitudinal coordinates}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
prediction_occ
}
\description{
A dataset containing geo-referenced occurrences of 100 Orchid species without
existing IUCN Red List assessment ("Not Evaluated"). This is example data
to predict the IUCN status.
}
\keyword{datasets}
