% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_labels.R
\name{prepare_labels}
\alias{prepare_labels}
\title{Format IUCN Red List categories for IUCNN}
\usage{
prepare_labels(
  x,
  species = "species",
  labels = "labels",
  accepted_labels = c("CR", "EN", "VU", "NT", "LC"),
  level = "detail",
  threatened = c("CR", "EN", "VU")
)
}
\arguments{
\item{x}{a data.frame or a list. If a data.frame, two columns with the species names and IUCN categories
respectively. THe column naems are defined by the species and labels arguments. If a list, expecting
the format as returned by \link[rredlist]{rl_search}.}

\item{species}{a character string. The name of the column with the species names.}

\item{labels}{a character string. The name of the column with the labels (assessment categories).}

\item{accepted_labels}{a character string. The labels to be converted in to numeric values.
Entries with labels not mentioned (e.g. "DD") will be removed. The numeric labels returned by the
function will correspond to the order in this argument.
For instance with default settings, "CR ->0, LC -> 4.}

\item{level}{a character string. The level of output level detail. IF "detail"
full iucn categories, if "broad" then 0 = Threatened, and 1 = NOt threatened.}

\item{threatened}{a character string. Only if level=="broad", Which labels to consider threatened.}
}
\value{
a data.frame with species names and numeric labels
}
\description{
Converting IUCN category labels into numeric categories required by \code{\link{train_iucnn}}.
}
\note{
See \code{vignette("Approximate_IUCN_Red_List_assessments_with_IUCNN")} for a
tutorial on how to run IUCNN.
}
\examples{
data("training_labels")
prepare_labels(training_labels)

}
