% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training_labels}
\alias{training_labels}
\title{IUCN threat categories for 884 orchid species}
\format{
A data frame with 889 rows and 2 variables:
\describe{
\item{species}{The canonical species name}
\item{category_detail}{The IUCN conservation assessment, converted to numerical for use with IUCNN.
0 =  Critically Endangered (CR), 1 = Endangered (EN), 2 = Vulnerable (VU), 3 = Near Threatened (NT), 4 = Least Concern (LC)}
}
}
\source{
\url{https://www.iucnredlist.org/}
}
\usage{
training_labels
}
\description{
A dataset containing the International Union for the Conservation of Nature's Global
Red List conservation assessments for 884 species of orchids.
}
\keyword{datasets}
