% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_numeric_labels.R
\name{translate_numeric_labels}
\alias{translate_numeric_labels}
\title{Translate numeric labels predicted by trained model into IUCN conservation status}
\usage{
translate_numeric_labels(
  predicted_labels,
  label_order = c("LC", "NT", "VU", "EN", "CR")
)
}
\description{
Translate numeric labels predicted by trained model into IUCN conservation status
}
