% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{feature_importance}
\alias{feature_importance}
\title{Evaluate importance of training features}
\usage{
feature_importance(
  input_model,
  n_permutations = 100,
  predictions_outdir = "",
  verbose = FALSE,
  feature_blocks = "auto",
  provide_indices = FALSE,
  unlink_features_within_block = TRUE
)
}
\description{
Evaluate importance of training features
}
