% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_get_pdp.R
\name{iucnn_get_pdp}
\alias{iucnn_get_pdp}
\title{Get partial dependence probabilities for IUCN Categories}
\usage{
iucnn_get_pdp(
  x,
  dropout_reps,
  feature_blocks = list(),
  provide_indices = FALSE
)
}
\arguments{
\item{x}{iucnn_model object, as produced as output
when running \code{\link{iucnn_train_model}}}

\item{dropout_reps}{integer, (default = 100). The number of how often the
predictions are to be repeated (only for dropout models). A value of 100 is
recommended to capture the stochasticity of the predictions, lower values
speed up the prediction time.}

\item{feature_blocks}{a list.}

\item{provide_indices}{logical. Set to TRUE if custom \code{feature_blocks}
are provided as indices. Default is FALSE.}
}
\description{
Get partial dependence probabilities for IUCN Categories
}
