% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn_prepare_phy.R
\name{iucnn_prepare_phy}
\alias{iucnn_prepare_phy}
\title{Prepare Phylogenetic Eigenvectors to Extract Phylogenetic Features}
\usage{
iucnn_prepare_phy(phy, variance_fraction = 0.8, numeigen = NULL)
}
\arguments{
\item{variance_fraction}{select number of eigenvalues that cumulative explain a specified fraction of variance}

\item{numeigen}{select the number eigenvalues}
}
\value{
a matrix of phylogenetic Eigenvectors
}
\description{
Extract Phylogenetic Eigenvectors to represent the phylogeentic relationsships
in a matrix format ot use as NN features.
}
\examples{
library(ape)
tree <- rphylo(n = 10, birth=0.1, death=0)
iucnn_prepare_phy(phy = tree)

}
\seealso{
Other Feature preparation: 
\code{\link{iucnn_bias_features}()},
\code{\link{iucnn_biome_features}()},
\code{\link{iucnn_climate_features}()},
\code{\link{iucnn_cnn_features}()},
\code{\link{iucnn_footprint_features}()},
\code{\link{iucnn_geography_features}()},
\code{\link{iucnn_phylogenetic_features}()},
\code{\link{iucnn_prepare_features}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
