% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pdp.R
\name{plot.iucnn_pdp}
\alias{plot.iucnn_pdp}
\title{Plot partial dependence probabilities}
\usage{
\method{plot}{iucnn_pdp}(x, ask = FALSE, features = NULL, uncertainty = TRUE, col = NULL, ...)
}
\description{
Plot partial dependence probabilities
}
