% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_iucnn.R
\name{predict_iucnn}
\alias{predict_iucnn}
\title{Predict IUCN Categories from Features}
\usage{
predict_iucnn(x, model, target_acc = 0, return_IUCN = TRUE)
}
\arguments{
\item{x}{a data.set, containing a column "species" with the species names, and
subsequent columns with different features,
in the same order as used for \code{\link{train_iucnn}}}

\item{model}{the information on the NN model returned by \code{\link{train_iucnn}}}

\item{target_acc}{numerical, 0-1. The target accuracy of the overall model.
Species that cannot be classified with
enough certainty to reach this accuracy are classified as DD (Data Deficient).}

\item{return_IUCN}{logical. If TRUE the predicted labels are translated
into the original labels.
If FALSE numeric labels as used by the model are returned}
}
\value{
outputs an \code{iucnn_predictions} object containing the predicted
labels for the input species.
}
\description{
Uses a model generated with \code{\link{train_iucnn}}
to predict the IUCN status of
Not Evaluated or Data Deficient species based on features, generated
from species occurrence records with \code{\link{prep_features}}.
These features should be of the same type as those used for training the
model.
}
\note{
See \code{vignette("Approximate_IUCN_Red_List_assessments_with_IUCNN")} for a
tutorial on how to run IUCNN.
}
\examples{
\dontrun{
data("training_occ") #geographic occurrences of species with IUCN assessment
data("training_labels")# the corresponding IUCN assessments
data("prediction_occ") #occurrences from Not Evaluated species to prdict

# 1. Feature and label preparation
features <- prep_features(training_occ) # Training features
labels_train <- prep_labels(training_labels) # Training labels
features_predict <- prep_features(prediction_occ) # Prediction features

# 2. Model training
m1 <- train_iucnn(x = features, lab = labels_train)

# 3. Prediction
predict_iucnn(x = features_predict,
             model = m1)
}


}
