% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_features.R
\name{prep_features}
\alias{prep_features}
\title{Prepare features for an IUCNN model}
\usage{
prep_features(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  type = c("geographic", "biomes", "climate", "humanfootprint")
)
}
\arguments{
\item{x}{a data.frame of species occurrence records including three columns with
species name, longitudinal coordinates and latitudinal coordinates (both decimal).}

\item{species}{a character string. The name of the column with the species names.}

\item{lon}{a character string. The name of the column with the longitude.}

\item{lat}{a character string. The name of the column with the latitude.}

\item{type}{character. The type of features to calculate. Possible options are
\dQuote{geographic}, \dQuote{biome}, \dQuote{climate},
\dQuote{human footprint}.
All except bias are the default.
"\dQuote{bias} is only recommended up to the regional scale or below.}
}
\value{
a data.frame of bias features
}
\description{
A wrapper function to prepare all default features included in IUCNN:
geographic, biomes, climate, human footprint.
If desired, bias features need to be calculated separately with ft_bias.
For more control over feature preparation, you can use the
\code{\link{ft_geo}}, \code{\link{ft_biom}}, \code{\link{ft_clim}},
\code{\link{ft_foot}} functions.
}
\details{
Without internet access, only geographic features are calculated,
if the sampbias package is not installed, the bias features are skipped.
}
\examples{
\dontrun{
dat <- data.frame(species = c("A","B"),
                  decimallongitude = runif (200,10,15),
                  decimallatitude = runif (200,-5,5))

prep_features(dat)
}

}
\seealso{
Other Feature preparation: 
\code{\link{ft_biom}()},
\code{\link{ft_clim}()},
\code{\link{ft_foot}()},
\code{\link{ft_geo}()}
}
\concept{Feature preparation}
\keyword{Feature}
\keyword{preparation}
